+++ example/minibitty/minibitty_exe_prv.h
@@ -88,6 +88,7 @@ int32_t app_match(struct app_s * self, const char * filter);
 typedef int (*command_fn)(struct app_s * self, int argc, char * argv[]);
 
 int on_adapter(struct app_s * self, int argc, char * argv[]);
+int on_cal(struct app_s * self, int argc, char * argv[]);
 int on_firmware(struct app_s * self, int argc, char * argv[]);
 int on_fwup(struct app_s * self, int argc, char * argv[]);
 int on_force_remove(struct app_s * self, int argc, char * argv[]);